+++ include/SearchGraphNode.hpp
@@ -5,19 +5,19 @@
 class SearchGraphNode{
 private:
     SearchGraphNode* parentNode;
-    unsigned int pathCostToNode;
-    float totalCost;
+    double pathCostToNode;
+    double totalCost;
     GraphNode* myGraphNode;
 public:
     SearchGraphNode();
     ~SearchGraphNode();
     void setParentNode(SearchGraphNode* newParentNode);
-    void setPathCost(const unsigned int pathCost);
-    void setTotalCost(float totalCost);
+    void setPathCost(const double pathCost);
+    void setTotalCost(double totalCost);
     void setGraphNode(GraphNode* myGraphNode);
     SearchGraphNode* getParentNode();
-    const unsigned int getPathCost();
-    const float getTotalCost();
+    const double getPathCost();
+    const double getTotalCost();
     GraphNode* getGraphNode();
     unsigned int getNodeId();
 };
